backend/requirements.txt
# FastAPI Backend for Physical AI & Humanoid Robotics Book
# Core dependencies for RAG chatbot, personalization, and translation

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & Vector Store
qdrant-client==1.7.3
asyncpg==0.29.0
psycopg2-binary>=2.9.0  # Relaxed version for Windows compatibility
sqlalchemy==2.0.25
alembic==1.13.1

# AI & LLM
openai==1.10.0
tiktoken==0.5.2

# Authentication (Better-Auth will be added later via JavaScript)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
httpx==0.26.0
python-json-logger==2.0.7
